Split into a client module
