+++ client/src/main/java/ru/funpay4j/client/FunPayClient.java
