+++ client/src/main/java/ru/funpay4j/client/exceptions/InvalidCsrfTokenOrPHPSESSIDException.java
