+++ client/src/main/java/ru/funpay4j/client/objects/game/ParsedPromoGame.java
