+++ client/src/main/java/ru/funpay4j/client/objects/lot/ParsedLotCounter.java
