+++ client/src/main/java/ru/funpay4j/client/objects/offer/ParsedOffer.java
