+++ client/src/main/java/ru/funpay4j/client/objects/user/ParsedPreviewSeller.java
