+++ client/src/main/java/ru/funpay4j/client/objects/user/ParsedSellerReview.java
