+++ client/src/main/java/ru/funpay4j/client/objects/user/ParsedUser.java
