+++ core/src/test/java/ru/funpay4j/core/AuthorizedFunPayExecutorTest.java
@@ -12,14 +12,13 @@
  * limitations under the License.
  */
 
-package ru.funpay4j;
+package ru.funpay4j.core;
 
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import ru.funpay4j.core.AuthorizedFunPayExecutor;
 import ru.funpay4j.core.commands.offer.*;
 import ru.funpay4j.core.commands.user.UpdateAvatar;
 import ru.funpay4j.client.exceptions.InvalidGoldenKeyException;
